requirements.txt
Flask==3.0.3
RPi.GPIO==0.7.1
